// app/Interfaces/Interfaces.tsx
export interface IntProjectData {
  title: string;
  miniature: string;
  description: string;
  code: string;
  site: string;
}

export interface IFormValues {
  firstName: string;
  lastName: string;
  email: string | undefined;
  phone: string | undefined;
  subject: string;
  message: string;
}